refactor: move MITRE ATT&CK constants to a dedicated file for better organization
