+++ plugins/wazuh-analysis/public/groups/threat-intelligence/apps/mitre-att&ck/mitre-att&ck-app.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { AppMountParameters } from 'opensearch-dashboards/public';
-import { MITRE_ATTACK_TITLE } from '../../constants';
+import { MITRE_ATTACK_TITLE } from './constants';
 
 interface MitreAttackAppProps {
   params: AppMountParameters;
